app/requirements.txt
kafka-python==2.0.2
pip==21.2.4
pymongo==4.1.1
pyparsing==3.0.4
pyrsistent==0.18.0
python-dateutil==2.8.2
setuptools==61.2.0
Faker==13.14.0